refactor: update utils imports to use browser-client, enhance logging, and improve data fetching and component structure
- Changed imports of createClient to use '@/utils/supabase/browser-client' for client-side
- Replaced console.error with logger.error for better logging consistency
- Added pagination and selected specific fields in server-side runs list for efficiency
- Memoized step sorting and grouping in flow run details to optimize rendering
- Removed unused event handlers in flow run provider for performance
- Simplified run page component to handle server-side data fetch and client updates
- Updated loading UI for better user experience during data fetch
- Enhanced error handling with dedicated error display component
- Added motion components file for UI animations (new file)
