+++ examples/playground/components/flow-run-details.tsx
@@ -218,26 +218,31 @@ export default function FlowRunDetails({
             <div>
               {runData.step_states &&
                 (() => {
-                  // Sort step_states directly by step.step_index
-                  const sortedStepStates = [...runData.step_states].sort(
-                    (a, b) => {
-                      const aIndex = a.step?.step_index || 0;
-                      const bIndex = b.step?.step_index || 0;
-                      return aIndex - bIndex;
-                    },
-                  );
+                  // Use memoized sorted step states to avoid sorting on every render
+                  const { sortedStepStates, parallelSteps, regularSteps } = useMemo(() => {
+                    // Sort step_states directly by step.step_index
+                    const sorted = [...runData.step_states].sort(
+                      (a, b) => {
+                        const aIndex = a.step?.step_index || 0;
+                        const bIndex = b.step?.step_index || 0;
+                        return aIndex - bIndex;
+                      },
+                    );
 
-                  // Group parallel steps based on flow definition
-                  // We're specifically looking for summary and tags steps that run in parallel
-                  const parallelStepSlugs = ['summary', 'tags'];
-                  const parallelSteps = sortedStepStates.filter((step) =>
-                    parallelStepSlugs.includes(step.step_slug),
-                  );
+                    // Group parallel steps based on flow definition
+                    // We're specifically looking for summary and tags steps that run in parallel
+                    const parallelStepSlugs = ['summary', 'tags'];
+                    const parallel = sorted.filter((step) =>
+                      parallelStepSlugs.includes(step.step_slug),
+                    );
 
-                  // Other steps will be displayed normally
-                  const regularSteps = sortedStepStates.filter(
-                    (step) => !parallelStepSlugs.includes(step.step_slug),
-                  );
+                    // Other steps will be displayed normally
+                    const regular = sorted.filter(
+                      (step) => !parallelStepSlugs.includes(step.step_slug),
+                    );
+                    
+                    return { sortedStepStates: sorted, parallelSteps: parallel, regularSteps: regular };
+                  }, [runData.step_states]);
 
                   // Function to render a step
                   const renderStep = (
