+++ examples/playground/components/header-auth.tsx
@@ -2,10 +2,11 @@
 
 import { signOutAction } from "@/app/actions";
 import { hasEnvVars } from "@/utils/supabase/check-env-vars";
+import { logger } from "@/utils/utils";
 import Link from "next/link";
 import { Badge } from "./ui/badge";
 import { Button } from "./ui/button";
-import { createClient } from "@/utils/supabase/client";
+import { createClient } from "@/utils/supabase/browser-client";
 import { SpinnerWrapper } from "./spinner-wrapper";
 import { useEffect, useState } from "react";
 
@@ -18,15 +19,15 @@ export default function AuthButton() {
     
     // Initial auth check
     supabase.auth.getUser().then(({ data, error }) => {
-      console.log('Auth check - user:', data.user);
-      console.log('Auth check - error:', error);
+      logger.log('Auth check - user:', data.user);
+      logger.log('Auth check - error:', error);
       setUser(data.user ?? null);
       setChecking(false);
     });
     
     // Listen for auth state changes
     const { data: { subscription } } = supabase.auth.onAuthStateChange((_event, session) => {
-      console.log('Auth state changed:', _event, session?.user);
+      logger.log('Auth state changed:', _event, session?.user);
       setUser(session?.user ?? null);
     });
